+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/google-vertex
 
+## 4.0.0-beta.125
+
+### Patch Changes
+
+- 0ad470b: feat(provider/google): add enterpriseWebSearch tool
+- Updated dependencies [0ad470b]
+  - @ai-sdk/google@3.0.0-beta.81
+
+## 4.0.0-beta.124
+
+### Patch Changes
+
+- Updated dependencies [d129d89]
+  - @ai-sdk/anthropic@3.0.0-beta.91
+
 ## 4.0.0-beta.123
 
 ### Patch Changes